+++ .gitignore
@@ -7,6 +7,7 @@ temp/*
 config/temp/*
 xpaths
 checksum
+**/package-lock.json
 
 # Runtime data
 pids
